EY-1795: Synleggjer intensjon ved å trekkje ut alt av tidspunkt-konverteringsfunksjonar til ei eiga fil
